% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeconnect.R
\name{get_connect}
\alias{get_connect}
\title{Calculate connectivity index}
\usage{
get_connect(lsc, alpha, cell_size)
}
\arguments{
\item{lsc}{Landscape raster with habitat and matrix cells. Habitat cells are
marked by integer number and matrix cells by NA values. The habitat values
are considered as fragment identities, i.e. each contagious fragments
should be labelled with a unique ID number. The lsc argument
can be either a integer matrix of a \code{\link[sp]{SpatialGridDataFrame}}}

\item{alpha}{Parameter that specifies the decay of connectivity with distance.}

\item{cell_size}{Cell size of the input raster (side length in meter)}
}
\value{
Dataframe with three columns
#' \enumerate{
  \item Fragment ID
  \item Fragment area in ha (multiple of cell_size^2)
  \item Fragment-level connectivity index
}
}
\description{
Calculate a connectivity index from a binary habitat raster, which considers
the distances among among the fragments, as well as the fragment areas
and shapes (May et al. 2013).
}
\examples{
library(sp)
lsc_raster <- read.asciigrid(system.file("extdata", "lachish_20m.txt", package = "shapeconnect"))
plot(lsc_raster)
connect_dat <- get_connect(lsc_raster, alpha = 50, cell_size = 10)

}
\references{
May et al. (2013). Metacommunity, mainland-island system or
island communities? - Assessing the regional dynamics of plant communities
in a fragmented landscape. Ecography, 36, 842–853.
}
